consolidate duplicate types
