+++ pkg/types/backend.go
@@ -428,63 +428,6 @@ type Runtime struct {
 	Gpus     []GpuType `json:"gpus"`
 }
 
-type GpuType string
-
-func (g *GpuType) UnmarshalJSON(data []byte) error {
-	var gpuStr string
-	err := json.Unmarshal(data, &gpuStr)
-	if err == nil {
-		*g = GpuType(gpuStr)
-		return nil
-	}
-
-	var gpuInt int
-	err = json.Unmarshal(data, &gpuInt)
-	if err != nil {
-		return err
-	}
-
-	if gpuInt == 0 {
-		*g = GpuType("")
-	} else if gpuInt > 0 {
-		*g = GpuType("T4")
-	}
-
-	return nil
-}
-
-func (g *GpuType) MarshalJSON() ([]byte, error) {
-	if *g == "" {
-		return []byte("0"), nil
-	}
-
-	return json.Marshal(string(*g))
-}
-
-func (g *GpuType) String() string {
-	return string(*g)
-}
-
-func GPUTypesFromString(gpu string) []GpuType {
-	gpus := []GpuType{}
-	gpuString := strings.Trim(gpu, " ")
-	if len(gpuString) > 0 {
-		for _, g := range strings.Split(gpuString, ",") {
-			gpus = append(gpus, GpuType(g))
-		}
-	}
-
-	return gpus
-}
-
-func GpuTypesToStrings(gpus []GpuType) []string {
-	var gpuStrings []string
-	for _, gpu := range gpus {
-		gpuStrings = append(gpuStrings, string(gpu))
-	}
-	return gpuStrings
-}
-
 // FilterFieldMapping represents a mapping between a client-provided field and
 // its corresponding database field, along with the values for filtering on the database field.
 type FilterFieldMapping struct {
